requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.43
Flask-Login==0.6.3
Flask-WTF==1.2.2
email-validator==2.3.0
Werkzeug==3.1.3
gunicorn==23.0.0
psycopg2-binary==2.9.10
requests==2.32.3
google-generativeai==0.8.5
google-cloud-speech==2.33.0
pydub==0.25.1
pandas==2.3.2
openpyxl==3.1.5
celery==5.3.6
redis==5.0.7
APScheduler==3.11.0
bcrypt==4.3.0
python-docx==1.2.0
python-telegram-bot==20.7
python-dotenv==1.0.0
instagrapi==2.1.3